Make quotes more ANSI compatible.
Using double quotes for values is MySQL specific.  Running against
a MySQL server with ANSI_QUOTES mode set results in an inability to
record migrations in the database as the migration name value is
treated as a column name.

This change makes the use of single quotes more consistent with all
existing database drivers and more compatible with ANSI usage.