Note about specifying `name` key with `connection`
There is currently no way to fetch database name from PDO connection
object and database name is required in order to check whether a
table exists in the database or not. If PDO connection object
is passed in database PHP configuration file, database name
in `name` key should be passed as well.

This documentation change explains how to use `connection` configuration
key along with `name` key.