Fixed Rollbacks
When rollbacking migrations, their order is now determined by when they were
executed (ie. their Start Time) as opposed to when they were created (ie. their
version name).

This means:
* With no parameters, the rollback command will rollback the last executed migration.
* With the -d (date) parameter, the migration start times are now used to
determine the target version to rollback to. In other words, all migrations
which were executed after that date will be rollbacked.
* With the -t (target version) parameter, all migrations which were executed
after that version was executed will be rollbacked.

In terms of testing:
* The testRollbacksByDate test was changed to be more "unit"y, ie. it now
asserts the rollback method is called with the right version (instead of
asserting the right version is rollbacked). A few test cases were also added to
it (via its provider).
* The testRollbacksByVersion was added.