Clean up composer dependencies and travis config
Having a `composer.lock` makes it harder stay current with dependencies,
so we remove it and prevent it from being added with `.gitignore`. This
was also locking in versions that caused Travis builds to fail.

Instead of explicitly testing Symfony 2.8 components, enable a "lowest
version" test on PHP 5.4 to prove that dependency versions are usable.

Move code coverage uploads into `after_script` so that a failed upload
will not cause the build to fail.

Finally, add detection for Postgres being available directly into the
adapter test and skip instead of failing. This simplifies the Travis
config and makes testing more robust.

Refs #732