Override default template source with command line options
The migration template can be created by migration template generator class or from a template file.
Either can be configured on the appropriate config file.

This commit inhibits the setting of both options.

In addition, the command line options can override the source, be it a class overriding a default template
file or a template file overriding a default class.

Unit tests included.