Added -s|--start-time option to Rollback command
When rollbacking, it is now possible to use the new option to order the
migrations by start time (ie. execution time) as opposed to the default
behaviour (where they are ordered by creation time).

This means:

- With only the -s|--start-time option, the rollback command will rollback
the last executed migration.
- With the -s|--start-time and -d (date) options, the migration start times
will be used to determine the target version to rollback to. In other words,
all migrations which were executed after that date will be rollbacked.
- With the -s|--start-time and -t (target version) options, all migrations
which were executed after the target version was executed will be
rollbacked.