Added support for *unsigned* custom `id`; added `signed` info in getColumns() & getPhinxType()
---

When creating custom id columns (e.g. `["id" => "custom_id"]`) the `signed` option got totally ignored. Example:

    // Note: it's important to set a custom name for the "id" column to reproduce this
    $this->table('foo', [ 'id' => 'custom_id', 'signed' => false ])->save();

This statement created a table called "foo" with an `id` column of type `int(11)` (signed!), which is obviously wrong. This commit fixes this behavior so the created column will be of type `int(11) unsigned`. Additionally I simplified the code a bit.

---

The method `MysqlAdapter::getPhinxType()` totally ignored the `signed` attribute, too. While writing unit tests for the fix mentioned above, I found out that it is not possible to check whetever a column's type is *really* signed or not, when reading the columns through `MysqlAdapter::getColumns()`. This is because

  a) MysqlAdapter::getColumns() doesn't set the signed attribute at all (e.g. with setSigned)
  b) MysqlAdapter::getPhinxType() doesn't return the information whetever the column's type is signed or not.

That's why this call:

    $this->adapter->getColumns('ntable')[0];

returned:

    class Phinx\Db\Table\Column{
        // ..
        protected $signed => bool(true) // (!)
        // ..
    }

Even if the column actually was an `int(11) unsigned` in the database.

To fix this, this commit added the `signed` property to the return value of `getPhinxType()` and getColumns() now makes use of it to provide accurate informations to the caller.