Start internals refactor
Using abumdant value objects to implement an interpreter-like machine
with the name "Plan".

The idea is that the plan can group the actions in a more intelligent
way, that is also faster for the database to execute in most cases.

Also refactored so that effects are treated more as data, this helps us
compose alter statamentes into a single call when it is an option.