Allow primary key change on non-integer columns
Problem exists when we want to change primary key of non-integer column. Match doesn't work so no primary key added. With this change if "creation" SQL is not with INTEGER then we parse all data from column, for example my_primary VARCHAR(100) NOT NULL will be changed to my_primary VARCHAR(100) PRIMARY KEY. (now it does not preg_match, would preg_match my_primary VARCHAR NOT NULL and change to INTEGER AUTOINCREMENT..)

Pseudo-reproduction:
1st migration:
$this->table('sometable', ['primary_key' => 'abc']);

2nd:
$this->table('sometable')->changePrimaryKey('dfg'); <- dfg is varchar

This is backwards compatible change if integers is passed without limitations and autoincrement is still there (probably it was intended?)