Fixed Command names
Symfony's Command require their name to return a string and don't allow a null
value, however in Phinx, in order to allow users to override the command names,
Commands in PhinxApplication are constructed without a name, and then in the
Command's configure method, a default name is set if getName is null.

Up until now, it worked as Symfony didn't enforce return types (types were
simply declared in PHPdoc), however for its version 5 Symfony plans to enforce
them, hence breaking Phinx commands.

To fix it, a COMMAND_NAME class constant has been introduced, and
PhinxApplication uses it when instantiating Commands.