SQLite will create a column as NULL by default
There is an error in SQLiteAdapter that creates the columns as NULL despite having indicated the field as NOT NULL. Currently, the only way to create a NOT NULL field is to specify a DEFAULT value for that column.
In order to allow for backward compatibility once the error is corrected, this change will cause
SQLiteAdapter continue to create a column as NULL by default when it is not specified whether it is null or not.