fix failing tests for mysql 8
MySQL 8.0.17 deprecated usage of the display width for integer
data types, stating that it did not really do anything as
it would allow integers to go beyond the specified width, and
was only really used when you wanted to pad a number. The
tests have been modified now to check on version of mysql to
determine what the expected type for integer columns will be,
but a more proper (and breaking) fix would be to probably not
bother with display widths at all for MysqlAdapter.

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>