Improve docs on what can and cannot be reversed
The docstring for the change function on migration is confusing
as it implies that there's specific methods you can call to
get reversible migrations, and that these methods are probably
part of the MigrationInterface class. However, the referred to
methods are actually part of the AdapterInterface, and are
not reversible if called directly, but rather you must go
through the Table API through its methods. This PR rewrites
the docs to make that a bit clearer by using plain english
for the actions that are reversible (with expectation that
end user will translate into Table API calls).

To also simplify keeping things up-to-date on what can and
cannot be reversed, and also making the docs a bit more of the
authoritative source, the docstring in the Migration template
is also simplified to directly point to the docs for a
description of what it does.

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>