Do not set precision for integer columns coming from postgres DB
The code would use the value of numeric_precision to set a limit
on integer columns. For integer values, this is set to the max
range of the type (e.g. 2^32 for integer), but is not actually
something the user can meaningfully use, nor has a counterpart
for info coming into postgres in a migration, so better to just
report null for limit instead.

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>