Allow name to be optional for create command
Currently, the create command requires a name to be provided,
and this is used for a suffix on the filename and then as the
class name. This PR makes the name optional, such that if
it is omitted, the system will automatically generate a name
that is valid. The filename is "<datetime>_migration.php" and
the class name is "V<Datetime>".

The filename is set-up such as the way it is since the
classname must start with a string, and the V is for a future
suggestion of potentially making phinx migration files be PSR-4
discoverable.

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>