Consolidate environment and environmentOptions printing for commands
Several commands were using the same sequence of checks and outputs
for getting the environment string, as well as then printing out
details about the environment. This commit consolidates all of that
repeated code into two functions within AbstractCommand. As part
of this, also fixed the bug of not being able to use sqlite
memory database without also setting a db name.

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>