Fix having to set name in config when using sqlite memory option
The phinx docs specify that you can do the following:

adapter: sqlite
memory: true

and it will give you the in-memory DB of sqlite to use. However,
before this commit, this would throw an error in migrate saying
that "name" was not set, requiring you to add "name" with a bogus
value to the above to get it work. This commit fixes it by having
it such that if memory option is set for sqlite, then just always
set name to :memory: regardless of what name was set to, retaining
BC and fixing the bug. Doing it this way also makes doing the
above equivalent to doing:

adapter: sqlite
name: :memory:

which is more inline with the general sqlite documentation on
how to think of the :memory: database name and phinx already
had proper logic set-up to handle those values correctly.

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>