Use ordinal_position for postgres getColumns
This ensures that getColumns will return results in column order
and not potentially row order in the information_schema.columns
table. While it's fairly impossible to get in this situation as
the columns table is a view of an underlying system table, better
to just follow best practices regardless in querying the table.

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>