fix using quoted and escaped values for mysql enums
The previous implementation was relatively simplistic in that
it would break if attempting to use a value with a ' character
in it and would give different result in getValues() after
pulling the column out of the DB. This patch makes the process
more robust by using PDO::quote to handle values going in properly
and then de-escapes things on the way out. Unfortunately, there's
no builtin method we can use to get "unescaped" strings and have
to do it by hand in custom fashion. Luckily, mysql's escape rules
on values coming out of the DB are relatively simple with just
really '' and \\ being the things we have to worry about.

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>