Fix adding columns to existing sqlite table with table constraints
If a table within sqlite had been made with table level constraints
(e.g. foreign key constraint), it would then throw an exception on
attempting to add any further columns. This is due to 16c83d1
appending the new column at the end of the create table SQL, after
any constraints. This patch modifies it so that it will inject the
new column after the last existing column, which will place it
before any constraints.

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>